add script to analyse participation

+++ R/concatenate_user_info.R
@@ -1,18 +1,24 @@
-if (FALSE){
+run = TRUE
 
-# read and give stats on VNE
-list_all_files <- paste0("data/",list.files("data", pattern = "5a9")) 
-data_all <- list_all_files |>
-  lapply(readr::read_csv) |>                                         # Combine data sets into one data set 
-  dplyr::bind_rows()
-
-if ("user_data.csv" %in% list.files("data")){
-  data_user <- read.csv("data/user_data.csv")
-  data_all <- dplyr::bind_rows(data_user, data_all)
+if (run){
+  
+  # read and give stats on VNE
+  files <- list.files("data/data_users_temp/", pattern = "5a9")
+  if(length(files ) > 0){
+    list_all_files <- paste0("data/data_users_temp/",) 
+    data_all <- list_all_files |>
+      lapply(readr::read_csv) |>                                         # Combine data sets into one data set 
+      dplyr::bind_rows()
+    
+    if ("user_data.csv" %in% list.files("data")){
+      data_user <- read.csv("data/user_data.csv")
+      data_all <- dplyr::bind_rows(data_user, data_all)
+    }
+    
+    
+    
+    readr::write_csv(data_all, "data/user_data.csv")
+    file.remove(list_all_files)
+  }
 }
 
-
-
-readr::write_csv(data_all, "data/user_data.csv")
-file.remove(list_all_files)
-}
